+++ projects/ce-code-editor/src/lib/code-editor/index.ts
@@ -1,0 +1,4 @@
+export * from './code-editor-header';
+export * from './code-editor-content';
+export * from './code-editor.component';
+export * from './code-editor.service';