+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # CeCodeEditor Changelog
 
+## [19.1.1]
+ - Add form dialog editor
+ - Refacto : split code into editor-header and editor-content
+
 ## [19.1.0]
  - Form object block compatibility